// admin/src/api/generated/models/UserPreferencesResponse.ts
export interface UserPreferencesResponse {
    /**
     *
     * @type {boolean}
     * @memberof UserPreferencesResponse
     */
    followLogs: boolean;
    /**
     *
     * @type {boolean}
     * @memberof UserPreferencesResponse
     */
    notifyDeploymentCompleted: boolean;
    /**
     *
     * @type {boolean}
     * @memberof UserPreferencesResponse
     */
    notifyDeploymentFailed: boolean;
    /**
     *
     * @type {boolean}
     * @memberof UserPreferencesResponse
     */
    notifyNodeUnhealthy: boolean;
    /**
     *
     * @type {string}
     * @memberof UserPreferencesResponse
     */
    timeFormat: string;
    /**
     *
     * @type {number}
     * @memberof UserPreferencesResponse
     */
    version: number;
}